bridge: split up processor.go
No code changes except for `break` -> `return`.

[ghstack-poisoned]
